+++ include/dsn/utility/utils.h
@@ -52,33 +52,6 @@ std::shared_ptr<T> make_shared_array(size_t size)
     return std::shared_ptr<T>(new T[size], std::default_delete<T[]>());
 }
 
-void time_ms_to_string(uint64_t ts_ms, char *str); // yyyy-MM-dd hh:mm:ss.SSS
-
-void time_ms_to_date(uint64_t ts_ms, char *str, int len); // yyyy-MM-dd
-
-void time_ms_to_date_time(uint64_t ts_ms, char *str, int len); // yyyy-MM-dd hh:mm:ss
-
-void time_ms_to_date_time(uint64_t ts_ms,
-                          int32_t &hour,
-                          int32_t &min,
-                          int32_t &sec); // time to hour, min, sec
-
-uint64_t get_current_physical_time_ns();
-
-// get unix timestamp of today's zero o'clock.
-// eg. `1525881600` returned when called on May 10, 2018, CST
-int64_t get_unix_sec_today_midnight();
-
-// `hh:mm` (range in [00:00, 23:59]) to seconds since 00:00:00
-// eg. `01:00` => `3600`
-// Return: -1 when invalid
-int hh_mm_to_seconds(dsn::string_view hhmm);
-
-// local time `hh:mm` to unix timestamp.
-// eg. `18:10` => `1525947000` when called on May 10, 2018, CST
-// Return: -1 when invalid
-int64_t hh_mm_today_to_unix_sec(string_view hhmm_of_day);
-
 // get host name from ip series
 // if can't get a hostname from ip(maybe no hostname or other errors), return false, and
 // hostname_result will be invalid value
