+++ src/core/core/env_provider.cpp
@@ -37,6 +37,7 @@
 #include <dsn/utility/utils.h>
 #include <dsn/c/api_utilities.h>
 #include <chrono>
+#include <dsn/utility/clock.h>
 
 namespace dsn {
