+++ src/core/perf_counter/perf_counters.cpp
@@ -36,6 +36,7 @@
 #include <dsn/tool-api/command_manager.h>
 #include <dsn/tool-api/task.h>
 #include <dsn/utility/string_view.h>
+#include <dsn/utility/clock.h>
 
 #include "perf_counter_atomic.h"
 #include "builtin_counters.h"
